calculate limitValue. Change frequency from float to int

+++ app/src/main/java/gemeenterotterdam/trillingmeterapp/HardwareSensor.java
@@ -60,7 +60,7 @@ public abstract class HardwareSensor implements SensorEventListener, HardwareUpd
     }
 
     @Override
-    public void updateFrequencyCounter(float[] maxFrequency) {
+    public void updateFrequencyCounter(int[] maxFrequency) {
         this.activity.updateFrequencyData(maxFrequency);
     }
 }
